+++ app/src/main/kotlin/com/pitchedapps/frost/fragments/FragmentContract.kt
@@ -16,6 +16,13 @@ interface FragmentContract : FrostContentContainer {
     val content: FrostContentParent?
 
     /**
+     * Defines whether the fragment is valid in the viewpager
+     * Or if it needs to be recreated
+     * May be called from any thread to toggle status
+     */
+    var valid: Boolean
+
+    /**
      * Helper to retrieve the core from [content]
      */
     val core: FrostContentCore?