+++ app/src/main/kotlin/com/pitchedapps/frost/parsers/NotifParser.kt
@@ -64,7 +64,8 @@ private class NotifParserImpl : FrostParserBase<FrostNotifs>(false) {
 
     override fun parseImpl(doc: Document): FrostNotifs? {
         val notificationList = doc.getElementById("notifications_list") ?: return null
-        val notifications = notificationList.getElementsByAttributeValueContaining("id", "list_notif_")
+        val notifications = notificationList
+                .getElementsByAttributeValueContaining("id", "list_notif_")
                 .mapNotNull(this::parseNotif)
         val seeMore = parseLink(doc.getElementsByAttributeValue("href", "/notifications.php?more").first())
         return FrostNotifs(notifications, seeMore)